Cosmonarchy Graphics Support
- General support for cosmonarchy data loading, cosmonarchy anims & cosmonarchy graphics
- Fix a bug where multi-(de)selections from the unit and sprite properties window were processed as individual actions (causing lag e.g. for selecting a single unit after select all for 1k units)
- Prevent sprites from auto-restarting iscripts after encountering an overflow (causing log spam/lag for invalid modded iscripts)
- Turn several unit/sprite related constants into calls getting the number of types loaded
- Use the loaded display names (which may be defaultDisplayNames or altered by mods) rather than using defaultDisplayNames directly
- Support for cosmonarchy unit names
- Support for cosmonarchy-modded iscripts (based on the iScriptId hiword)
- Use u32 iscript ids rather than size_t's (allowing for the aformentioned hiword extraction)
- Support for u16 flingy ids

+++ src/mapping_core/render/map_image.h
@@ -87,7 +87,7 @@ struct Animator
 
     void setActorDirection(u8 direction);
 
-    void initializeImage(size_t iScriptId);
+    void initializeImage(u32 iScriptId);
     void endImage();
 
     void createOverlay(u16 imageId, s8 x, s8 y, bool above);
